+++ README.md
@@ -2,9 +2,9 @@
 - [x] 토큰 발급하는 API 생성
 - [x] 내 정보 조회하기
   - [x] 토큰을 이용하여 본인 정보 응답하기 
-- [ ] 예약하기 개선 
+- [x] 예약하기 개선 
   - [x] 로그인된 사용자는 예약이 가능하다 
-  - [ ] 비로그인 사용자는 예약이 불가능하다 -> HTTP CODE 401
+  - [x] 비로그인 사용자는 예약이 불가능하다 -> HTTP CODE 401
 - [ ] 예약취소 개선
   - [ ] 자신의 예약인 경우 예약 취소가 가능하다. 
   - [ ] 자신의 예약이 아닌 경우 예약 취소가 불가능하다.
