+++ src/main/java/nextstep/auth/AuthorizationExtractor.java
@@ -1,5 +1,7 @@
 package nextstep.auth;
 
+import nextstep.support.AuthorizationException;
+
 import javax.servlet.http.HttpServletRequest;
 import java.util.Enumeration;
